requirements.txt
discord.py==2.3.2
python-dotenv==1.0.0
requests==2.31.0
aiohttp==3.9.1
python-dateutil==2.8.2
apify-client==1.6.1
pytest==8.0.0
pytest-asyncio==0.23.5 